requirements.txt
streamlit
python-dotenv
crewai==0.22.5
langchain==0.1.20
langchain-core==0.1.53
langchain-google-genai==0.0.8
google-generativeai==0.3.2
